Fix turning off notes at end of program / Ctrl+C
Now that the notes have timestamps, the following happens: the midi
buffer is filled with a lot of midi notes. Before this commit, the
turn-all-notes-off message was added to the end of that buffer. Afterward
adding that message, the program was exited and with that the buffer
was closed. Hence, the turn-all-notes-off message was never read.

That's why this commit closes the device and opens a new one to restart
buffer before sending the turn-all-notes-off message.

Co-authored-by: Enguerrand Decorne <decorne.en@gmail.com>